Cater for new way of recording rural addresses
Jane from Wangaratta described the requirement:

We are in the processes of  changing T1 Property to record rural address
information as an attribute of the address rather than a customer field
on the property.  This is so we can correctly and accurately record a
secondary address as a rural address and not just assume it is or is
not.