Add lpapabl table to Pathway extract
This table may contain the parcel's block number attribute (as
discovered at Nillumbik). This update extracts the new table at the last
step in the task so that if the table is not available at a particular
site, it would only crash after all other steps have completed. If no
sites complain about any crashing, then we can investigate updating the
parcel query (like at Nillumbik) to use the table for populating the
block value.