Add new filter for P edits
Exclude properties being matched to an approved parcel if the number is
already matched to a proposed parcel.

Based on common Monash scenario, but should be applicable to other
councils.

May require update to E edit query to take into account that invalid
property numbers will no longer be replaced due to this change.