P edits: include parcel pfi for proposed parcels
This addresses an issue raised by Port Phillip:

DELWP have changed the way they transfer ‘proposed parcels’ to ‘active’ which invalidates some of their previous assumptions.  Basically, as I read it, if a ‘proposed’ series of parcels become ‘active’ before the ‘propsed’ properties are populated with a valid propnum, the newly activated parcels become linked to both the old (existing) properties AND to the still-to-be-activated (still proposed) properties.  This means the only way to update the ‘proposed’ properties is to use the PFIs – using the parcel plan/lot description fails saying the parcel is associated to multiple properties.  Where there are multiple parcels linked to multiple proposed properties, we have to use the PFIs for the appropriate linked parcel and property or we risk sending the incorrect address to Land Registry – at least that is my reading of it.